feat: restructure PriceChart component, remove redundant updates of the cards while selecting an active one
