+++ frontend/src/components/PriceChart/PriceChart.tsx
@@ -1,43 +1,29 @@
 import { useEffect } from 'react';
+import { useAppDispatch } from '@/hooks/redux';
+import { fetchPriceHistory } from '@/store/priceHistorySlice';
+import PriceChartInner from './src/PriceChartInner';
 import './priceChart.css';
-import { Line, LineChart, XAxis, YAxis, ResponsiveContainer } from 'recharts';
-import { useAppDispatch, useAppSelector } from '@/hooks/redux';
-import { fetchPriceHistory, selectors } from '@/store/priceHistorySlice';
-import Loading from '@/components/Loading';
+
 type PriceChartProps = {
   symbolId: string | null;
 };
 
 const PriceChart = ({ symbolId }: PriceChartProps) => {
   const dispatch = useAppDispatch();
+
   useEffect(() => {
     if (symbolId) {
       dispatch(fetchPriceHistory(symbolId));
     }
   }, [dispatch, symbolId]);
 
-  const apiState = useAppSelector(selectors.apiState);
-  const data = useAppSelector(selectors.selectPriceHistory);
-  const symbolInfo = useAppSelector(selectors.selectSymbolInfo);
-
-  if (apiState.loading && symbolId !== null)
-    return (
-      <div className="priceChart">
-        <Loading />
-      </div>
-    );
-  if (apiState.error) return <div className="priceChart">Failed to get price history!</div>;
-  if (!symbolId) return <div className="priceChart">Select stock</div>;
   return (
     <div className="priceChart">
-      <div>{symbolInfo}</div>
-      <ResponsiveContainer width="100%" height="100%">
-        <LineChart data={data.map((e) => ({ ...e, time: new Date(e.time).toLocaleTimeString() }))}>
-          <Line type="monotone" dataKey="price" stroke="#8884d8" dot={false} />
-          <XAxis dataKey="time" />
-          <YAxis />
-        </LineChart>
-      </ResponsiveContainer>
+      <h3 className="priceChart__title">Price history</h3>
+
+      {symbolId && <PriceChartInner />}
+
+      {!symbolId && <p>Select stock</p>}
     </div>
   );
 };
