+++ frontend/src/components/PriceChart/priceChart.scss
@@ -1,0 +1,24 @@
+.priceChart {
+  width: 400px;
+
+  &__title {
+    text-transform: uppercase;
+  }
+
+  &__inner {
+    height: 300px;
+    display: flex;
+  }
+
+  @media (max-width: 1023px) {
+    width: 100%;
+
+    &__inner {
+      height: 180px;
+    }
+  }
+
+  .loading {
+    height: 200px;
+  }
+}
