+++ frontend/src/components/SymbolsView/src/stocks.css
@@ -1,11 +1,8 @@
 .stocks {
   height: 700px;
   margin: 1.5rem 0;
-  display: flex;
-  flex-direction: row-reverse;
 }
 .stocks__item--primary {
-  width: calc(100% - 420px);
   padding: 1rem;
   overflow-y: scroll;
   background: #e6e6e6;
@@ -15,9 +12,20 @@
   display: none;
 }
 .stocks__item--secondary {
-  width: 420px;
   padding: 0 1rem;
   box-sizing: border-box;
+}
+@media (min-width: 1024px) {
+  .stocks {
+    display: flex;
+    flex-direction: row-reverse;
+  }
+  .stocks__item--primary {
+    width: calc(100% - 420px);
+  }
+  .stocks__item--secondary {
+    width: 420px;
+  }
 }
 
 /*# sourceMappingURL=stocks.css.map */
