+++ frontend/src/components/SymbolsView/src/stocks.scss
@@ -4,12 +4,9 @@ $item-secondary-width: 420px;
 .stocks {
   height: $height;
   margin: 1.5rem 0;
-  display: flex;
-  flex-direction: row-reverse;
 
   &__item {
     &--primary {
-      width: calc(100% - #{$item-secondary-width});
       padding: 1rem;
       overflow-y: scroll;
       background: #e6e6e6;
@@ -21,9 +18,22 @@ $item-secondary-width: 420px;
     }
 
     &--secondary {
-      width: $item-secondary-width;
       padding: 0 1rem;
       box-sizing: border-box;
+    }
+  }
+
+  @media (min-width: 1024px) {
+    display: flex;
+    flex-direction: row-reverse;
+
+    &__item {
+      &--primary {
+        width: calc(100% - #{$item-secondary-width});
+      }
+      &--secondary {
+        width: $item-secondary-width;
+      }
     }
   }
 }
